Dynamic group by (#116)
* [NEW] Added dynamic zip iterator to utils to iterate over an unknown at compile time number of iterators simultaneously.

* Use DynamicZip in group by to allow group-by by more than 5 fields.

WARNING: In order to achieve this the signature for `hash_join::prepare_hashed_relation`
and `group_by::groupby` have been modified and they do not constraint the type to be
copyable anymore. Then, move can occur when using this function. This change is needed as
DynamicZip returns a Vec and not a tuple as Zip do.

* Added tests for group by 6 elements.

* Added static group by calculation to up to 12 fields (maximum size for tuples which implement the Copy trait).
Use dynamic group by for grouping more than 12 fields, that means that for more than 12 fields heap allocation is
used instead.

* Modified test to cover corner cases.
- Group by static by 12 columns (maximum static group by)
- Group by dynamic by 13 columns (minimum dynamic group by)

* Removed `println!` added during testing phase in `groupby` function.