Utf8 no null multiple chunk parallel iterator (#159)
* Add parallel iterators as a submodule of iterators.
The goal of this commit is to be able to access the sequential iterators private fields
from the parallel iterator module.

This is because a parallel iterator depends on several phases:
- Split the iterator list in smaller partial iterators, using a producer.
- Asign a thread to each iterator.
- Convert the partial iterator in each thread in a sequencial iterator, and iterate
  it sequencially. This last part is the reason why we want to be able to access
  the private fields of sequential iterators, to compose sequential iterators with
  an offset and a size relative to the thread. Reusing the sequential iterator code.

* Add a parallel feature to cargo to allow the use of parallel iterators.

* Added parallel iterator implementation for:
- Utf8ParIterSingleChunk
- Utf8ParIterSingleChunkNullCheck
- Utf8ParIterManyChunk
- Utf8ParIterManyChunkNullCheck

* Added macro to generate parallel iterators code. This macro requires the sequential iterators
to implement the From<Producer> trait in order to be used.

* Implemented traits `ExactSizeIterator` and `DoubleEndedIterator` to `Utf8IterCont`.
Created parallel iterator for `Utf8IterCont`.

* Implement `Utf8ParChunkIterDispatch` to provide static polymorphism to parallel iterators.
Implement `IntoParallelIterator` for `&'a Utf8Chunked` which will choose the implementation
of `ParallelIterator` that is more efficient for the current `&'a Utf8Chunked`:
- If the chunked array has only one chunk and no null element then use `Utf8ParIterSingleChunk`.
- If the chunked array has only one chunk and does have null element then use `Utf8ParIterSingleChunkNullCheck`.
- If the chunked array has many chunks and no null element then use `Utf8ParIterManyChunk`.
- If the chunked array has many chunks and does have null element then use `Utf8ParIterManyChunkNullCheck`.

* Implemented `Utf8IterContManyChunk` for non nullable iterators with many chunks.
Implemented dispatcher for non-null iterators with a variant for single chunk and
other for many chunks.
Implement a `IntoNoNullIter` which get the most efficient implementation of the sequential
iterator for a non-nullable chunked array.

* Format with `cargo fmt`.

* Fixed `Utf8IterContManyChunk` constructor.
Implement `ExactSizeIterator` for `Utf8IterContManyChunk`.

* Added `Utf8ParIterContManyChunk` parallel iterator for many chunks non-nullable chunked arrays.
Added parallel dispatcher for NoNull<&'a Utf8Chunked> which return non-nullable parallel iterators.
Implement `IntoParallelIterator` for `NoNull<&'a Utf8Chunked>` which returns the most efficient
parallel iterator for non-nullable `&'a Utf8Chunked`. Depending of the number of chunks of the
chunked array.

* Format code with `cargo fmt`.

* Added tests for `Utf8IterCont` and `Utf8IterContManyChunk`.

* Added tests for parallel iterators. The test do the following operations: map, filter and fold; over parallel and sequential
iterators and compares that the result are the same.

Fixed many chunk parallel iterators, whose right index was wrong. Created helper function to calculate the right chunk and the
index in the right chunk by:
1. Compute the `chunk_idx` and the `remainder_idx` using `index_to_chunked_index`, for the `right_index`.
2. If the `right_index` is not 0 and the `remainder_idx` is 0, then shift the index to point the previous
chunk at position `chunk.len()`.

* Refactor iterators names to differentiate iterators which return option and those which
return the unwrapped value, previously called cont iterators.
Also, added new documentation to explain the different iterators.

* Refactor test names to match the new iterator names.

* remove macros to reduce code bloat; #157

* Make take work for Object types

* Add parallel iterators as a submodule of iterators.
The goal of this commit is to be able to access the sequential iterators private fields
from the parallel iterator module.

This is because a parallel iterator depends on several phases:
- Split the iterator list in smaller partial iterators, using a producer.
- Asign a thread to each iterator.
- Convert the partial iterator in each thread in a sequencial iterator, and iterate
  it sequencially. This last part is the reason why we want to be able to access
  the private fields of sequential iterators, to compose sequential iterators with
  an offset and a size relative to the thread. Reusing the sequential iterator code.

* Add a parallel feature to cargo to allow the use of parallel iterators.

* Added parallel iterator implementation for:
- Utf8ParIterSingleChunk
- Utf8ParIterSingleChunkNullCheck
- Utf8ParIterManyChunk
- Utf8ParIterManyChunkNullCheck

* Added macro to generate parallel iterators code. This macro requires the sequential iterators
to implement the From<Producer> trait in order to be used.

* Implemented traits `ExactSizeIterator` and `DoubleEndedIterator` to `Utf8IterCont`.
Created parallel iterator for `Utf8IterCont`.

* Implement `Utf8ParChunkIterDispatch` to provide static polymorphism to parallel iterators.
Implement `IntoParallelIterator` for `&'a Utf8Chunked` which will choose the implementation
of `ParallelIterator` that is more efficient for the current `&'a Utf8Chunked`:
- If the chunked array has only one chunk and no null element then use `Utf8ParIterSingleChunk`.
- If the chunked array has only one chunk and does have null element then use `Utf8ParIterSingleChunkNullCheck`.
- If the chunked array has many chunks and no null element then use `Utf8ParIterManyChunk`.
- If the chunked array has many chunks and does have null element then use `Utf8ParIterManyChunkNullCheck`.

* Implemented `Utf8IterContManyChunk` for non nullable iterators with many chunks.
Implemented dispatcher for non-null iterators with a variant for single chunk and
other for many chunks.
Implement a `IntoNoNullIter` which get the most efficient implementation of the sequential
iterator for a non-nullable chunked array.

* Format with `cargo fmt`.

* Fixed `Utf8IterContManyChunk` constructor.
Implement `ExactSizeIterator` for `Utf8IterContManyChunk`.

* Added `Utf8ParIterContManyChunk` parallel iterator for many chunks non-nullable chunked arrays.
Added parallel dispatcher for NoNull<&'a Utf8Chunked> which return non-nullable parallel iterators.
Implement `IntoParallelIterator` for `NoNull<&'a Utf8Chunked>` which returns the most efficient
parallel iterator for non-nullable `&'a Utf8Chunked`. Depending of the number of chunks of the
chunked array.

* Format code with `cargo fmt`.

* Added tests for `Utf8IterCont` and `Utf8IterContManyChunk`.

* Added tests for parallel iterators. The test do the following operations: map, filter and fold; over parallel and sequential
iterators and compares that the result are the same.

Fixed many chunk parallel iterators, whose right index was wrong. Created helper function to calculate the right chunk and the
index in the right chunk by:
1. Compute the `chunk_idx` and the `remainder_idx` using `index_to_chunked_index`, for the `right_index`.
2. If the `right_index` is not 0 and the `remainder_idx` is 0, then shift the index to point the previous
chunk at position `chunk.len()`.

* Refactor iterators names to differentiate iterators which return option and those which
return the unwrapped value, previously called cont iterators.
Also, added new documentation to explain the different iterators.

* Refactor test names to match the new iterator names.

Co-authored-by: Ritchie Vink <ritchie46@gmail.com>