Refactored iterators. The following tasks have been carried out: 1. For every chunked array leave only 4 iterators.     a. IteratorSingleChunk -> No null values, only one chunk, returns the value `T` unwrapped.     b. IteratorSingleChunkNullCheck -> Null values, only one chunk, returns `Option<T>`.     c. IteratorManyChunk -> No null values, many chunk, returns the value `T` unwrapped.     d. IteratorManyChunkNullCheck -> Null values, many chunk, returns `Option<T>`.
2. Included `SomeIterator` newtype. This newtipe is a wrapper around the iterators which returns
`T` and convert then into iterators which returns `Option<T>`, without runtime additional cost.
It is useful for `IteratorSingleChunk` and `IteratorManyChunk`, to return the same value as
`IteratorSingleChunkNullCheck` and `IteratorManyChunkNullCheck`, required to implement `IntoIterator` trait.

3. Created a trait `PolarsIterator`, for iterators over polar types. A `PolarIterator` requires
`ExactSizeIterator` and `DoubleEndedIterator` to be implemented.

4. Removed dispatcher static type, in favor of `dyn PolarsIterator`, as the former implementation
implies more code generation, and the optimization is not worth it, as it is only call once per a whole
chunked array. Dynamic dispatching is good in this situation.

5. Refactored macro generation to allow a function as an optional macro argument, and only use it if necessary.
This is important because the implementation of bool does not requires the execution of a function in every
return anymore.

6. Utf8Chunks are generated now with macros, having the exact same behaviour. This produces 400 lines of code
less than before.

7. Added more documentation comments to make clearer what each function and macro does. And what the different
structs are for.