shallow implementation of join prune
When a join is done on tables from the same src
the join can be dropped in specific cases.
This commit drops the join if the previous node
in the logical plan is an aggregation or a projection.
(on certain conditions)
A later commit will inspect, and combine lower nodes as
well. #449