Add support for reading gzipped CSV files with read_csv when use_pyarrow=False.
Add support for reading gzipped CSV files with read_csv when use_pyarrow=False.
Gzipped files are decompressed with igzip of python-isal, if it is
installed as it is around 30% faster than the gzip module of python.
Pyarrows CSV parser is still much faster for reading gzipped CSV files
(at least twice as fast as igzip).

Timings:
  Reading gzipped CSV file with pyarrow:
    CPU times: user 45.4 s, sys: 15.2 s, total: 1min
    Wall time: 25.4 s

  Reading gzipped CSV file with igzip.open() and DataFrame.read_csv():
    CPU times: user 45.1 s, sys: 20.9 s, total: 1min 5s
    Wall time: 1min 5s

  Reading gzipped CSV file with gzip.open() and DataFrame.read_csv():
    CPU times: user 59.3 s, sys: 24.7 s, total: 1min 23s
    Wall time: 1min 24s