chore[python]: Turn on mypy strict mode (#4415)
See the comment in pyproject.toml. It basically means we have not changed that much, as we still ignore the warning. But there is not really a feasible way until PyO3 supports generating type annotations. Otherwise, maintaining stubs would just be a huge waste of resources, as it is duplicating the "outer" python api.